+++ src/translate/Source.tsx
@@ -1,5 +1,14 @@
-import { Chip } from "@mui/material";
+import { Chip, IconButton } from "@mui/material";
+import ClearRoundedIcon from "@mui/icons-material/ClearRounded";
 import { Lang, TranslationResponse } from "./types";
+import styled from "styled-components";
+
+const IconContainer = styled.div`
+  position: absolute;
+  top: 0px;
+  right: 0px;
+  z-index: 99999;
+`;
 
 export const Source = ({
   question,
@@ -23,6 +32,15 @@ export const Source = ({
   return (
     <>
       <div className="autosize" data-replicated-value={question}>
+        <IconContainer>
+          <IconButton
+            onClick={() => setQuestion('')}
+            aria-label="copy"
+            size="small"
+          >
+            <ClearRoundedIcon fontSize="inherit" />
+          </IconButton>
+        </IconContainer>
         <textarea
           placeholder="Your text here..."
           value={question}